SCUBA2: rename to reflect how the primitive crops the image
The name has been updated to reflect the fact that this primitive
specifically uses an AST region to perform the cropping. The code has
also been tidied up a little and generalized to work with either Frame
or Group objects.

The output file name is returned via the primitive hash with the name
OUTFILE (even if specified on entry), rather than storing in the
object which should be done by the caller.

Warnigns are now issued if the map size is to be set to the minimum,
defined as 75 arcsec on a side.