Inst: allow non-QL/SUMMIT SCUBA-2 and ACSIS pipelines to run in batch mode today
If the UT date is the current date then it is assumed that the
pipeline is running "live". However, the live instances of the ACSIS
and SCUBA-2 pipelines use the recsuffix argument to specify custom
recipes, which means other instances of the SCUBA-2 and ACSIS
pipelines can safely be run in batch mode "today".