heterodyne: Cope with a new special case.
Some early spectra are entirely bad save some noisy edges, so would
pass through the not-all-bad test in this primitive.  Whereas later
we trim these edges, normally retaining the central 85%, and this
section would be bad.  So look for the SPECTRAL_TRIMMED internal header
and mimic what later primitives will do, and hence test for goodness
in the section actually used in recipes.