heterodyne: Preliminary version of the iterative flat field.
This combines whole observations and allows for the bias in the
comparisons by iterating to convergence.  Only initial tests this works
fine with the number of receptors changing and the degree of shift
reducing at each iteraiton.  However in about a quarter of cases the
code does not converge and is being investigated.

I commit it now because 2de9453dfc4ab5aa35cae068a63b00b9e9387740
included more than intended.  It was supposed to have just the missing
line from _REDUCE_SCIENCE_GRADIENT_STEER_ and none of the itersum
changes, which should be part of this commit.  Most of the excess was
innocuous, however _ITERATIVE_GROUP_PRODUCTION_ calls the new
primitive...