SCUBA2: major overhaul to now calculate peak and total flux of source
Previously this primitive calculated the fluxes of each source in a
SExtractor catalogue. However, it was never used in that form, and is
more useful as a simpler way of determining the peak and total flux of
a source.

Now the primitive performs aperture photometry to determine the total
flux, and fits the source to determine the peak flux. The associated
uncertainties are also derived and all values are returned in the
primitive hash with the following keys: TOTALFLUX, TOTALFLUXERR, PEAK,
PEAKERR, POSITION, and FITPAR.

The primitive may use a given image, or the current file in the
specified object or the file marked with a given tag.