SCUBA2: combine and use multiple images if present
This patch fixes a long-standing bug where multiple images for a given
SMU position were not actually being used in the focus cube.

Also increase the trimmed map size to 120" to allow for the source
being off-centre by a large amount (if a pointing was not done before
the focus).