heterodyne: Make more robust against bad values.
_QA_RMS_TSYS_CONSISTENCY_ was picking out the central frequency slice
of the VARIANCE array as reprsentative, but this overlooks the
possibility of this slice being msotly or all bad.  Since the
variances are allegedly independent of frequency, just collapse the
frequency axis taking the mean to give a representative plane.  We can
also trim at the same time saving the MANIC call.  I would like to
avoid the POW command too by making COLLAPSE use comp=error but this
exits with an error.

Remove the now unused temporary filename.