heterodyne: Allow for undefined TAU225ST header.
This may be fixed retrospectively in the data using the CSOfit values.
It could be handled in header translation, but this is not an ORAC_
header, so I decided for now to do it a simple way to permit more
processing jobs to complete at CADC after an rsync.  It is only accessed
in this primitive.