SCUBA2: find source position before creating cube
This primitive now coadds the images from each SMU position and
determines the offset of the source from the nominal 0,0 position
before creating the cube. This allows the shift to be applied to each
image before they are trimmed so the source should appear at the image
centre in the final cube.