lib: refresh next names list in orac_loop_flag
The orac_loop_flag subroutine previously prepared the list of
flag files for the next observation before entering the
while loop in which it waits for new data.  However for
SCUBA-2 this could happen before the meta file is written
indicating which subarrays are active and should have their
ok files waited for.  Then it would wait endlessly for ok
files which were never going to be created.  This patch
causes the list of flag files for the next observation
to be refreshed.  (This is done every time a dot is printed
so that it wouldn't necessarily be every time round the
while loop.)