SCUBA2: remove artifical (and unused) coadd mode
There's an easier way to tell if things are awry when making RUNNING
coadds. Rather than set an artificial coadd mode, check if the number
of members in the group is non-zero. If so, and there is no group file
then something weird has happened. Otherwise having no group file is
totally expected for the first pass through the primitive.