SCUBA2: ensure the Group contains the correct file if necessary
The "running" coadd is formed from the combination of the current
calibrated Frame image with the current Group image (which was formed
from the ... etc). In practice the coadd is tagged as the "reduced"
product so make sure the current file in the Group is that file. Of
course, this is not necessarily true in general but it is for the
current SCUBA-2 recipes.