heterodyne: Allow for poorly or undefined median in the overlap region.
Jan came across some data with entirely bad or mostly bad values in the
subband overlap regions leading to an undefined median.  Using a
default of zero will not be far off the mark.  If this is not done
the bad value offset is propgated along the sequence of subbands.