SCUBA2: updates to help deal with marginal images
Images with low S/N can occasionally be problematic and a fairly
common error at 450 um in poor weather was that the source position to
be fitted was either right at the edge of or not even on the
image. This was due to the fact that the search box was limited to 11
pixels *only* for calibrators, whereas it should always be limited for
pointing and focus observations. This caused problems because the
position was used as input for fitting a smaller version of that
image.

In addition, add a check for pointing/focus observations for the
distance between the peak S/N position and the centre and warn if it's
greater than 30". In this case the attempt to locate the source will
be abandoned.