heterodyne: Allow two recipes to work with DAS hybrid mode.
The focus recipe is effectively unchanged for ACSIS.  However, the
pointing recipe now has the noisy spectral ends removed.  Thus the
pointing estimates will be slightly different for the line emission,
and noticeably different, but improved, for the continuum pointing,
at a small cost of processing time.

The hybrid mode flag is set in the _REDUCE_SCIENCE_STEER_ primitive.