heterodyne: Offer a choice of methods to flat field.
The addition is for the simple summation of the signal between
velocity limits, expected to defined the line emission. It adds three
new arguments for the chosen method and velocity limits, which are
also recipe parameters.  Although the sum method does not require
iteration, I added it here because this primitive combines all the
subscans to improve the accuracy of relative fluxes by receptor and
makes maps for each receptor individually (both unlike
_REMOVE_FLAT_FIELD_).  It could be divided into two primitives if
that's deemed clearer.