heterodyne: Create missing VARIANCE component with the noise.
Instead of filling the VARIANCE with bad values, estimate it from
the smoothed noise.  The noise is determined from the detrended and
median-filtered spectra.  This does not address the erroneous Tsys
values.

Also put these steps in a loop to handle MAKECUBE creating tiles whose
names are not set to $out.