heterodyne: Protect against FINDBACK failing.
When FINDBACK fails, attempt an equivalent block smooth, where the
smoothing is necessary.  For the high-frequency interference
filtering, this is believed not to be required, where in effect we
skip the FINDBACK step.  The main peaks do stand out clearly for the
first filter.  For the ringing, a block smooth make incorporate the
signal being detected and after subtraction weaken the filter.  The
ringing filter is still not finalised so leave consideration of a
FINBACK alternative until further analysis and refinement of the
ringing filter.