heterodyne: Dynamically adjust the trimming to reain the minimum overlap.
There are observations where strong spectral lines lie in the overlap
region.  The default method of using MFITTREND to determine the
baseline ranges and compare these for overlap fails in such cases.
The alternative had been to use the supplied trimming percentages,
even if these removed all overlap.  Then current change reduces the
trim percentage to enable the minimum overlap in pixels.  This is far
from ideal as the DC correction can be misplaced because of the few
pixels in a strong and noisier (in absolute terms) region.  Hence
there is commentary made.  I dithered over making it a warning
as it might induce a Fault from Harriet when such instances are
reprocessed, and the offset baselines are evident.