lib: Add check for missing files with gz versions
Before converting input files, check whether each file is
missing but is present in gzipped form.  This allows
data to be reduced where the raw files have been gzipped
but the OK files still list the original file names.

When ORAC-DR ungzips files into ORAC_DATA_OUT it does not
unlink them at the end of the reduction.  This is also the
case with gzipped files specifed by with the -files
option.  Unfortunately there doesn't appear to be a way
for _convert_and_link_nofrm to mark them as temporary
using $Frm->tempraw because no frame is available to it.