heterodyne: Protect FINDCLUMPS against crazy artefactual values.
This addresses Fault 20130612.006.  While the large value that
FINDCLUMPS was invited to contour to separate clumps originated in
MFITTREND, it is a quicker fix than getting to the bottom of why one
edge baseline voxel was c. -dex9 (positive once subtracted).  Also
generating an abnormal value is always a possibility while fitting to
noisy data.  The chosen range should not exclude any known
astronomical signal and make the maximum number of contours in
FINDBACK more manageable.