heterodyne: Restrict to HARP.
I have dithered over this for a couple of days.  The real criterion is
the number of receptors so one can later assure that there will be
sufficient to form reliable statistics and delineate outliers.  I
first coded it that way but then it becomes messy with possibly
already rejected receptors for different frames.  So I went for the
JCMT-specific approach.