heterodyne: Have another go at handling anomalous data before moment-map creation.
Main thing is to check that the median rms is not zero as this is used
to set the contours in FINDCLUMPS.  Looking for low values eliminates
many good data---not sure what I was thinking there and can only
suspect that I must have mixed it with $medianrms.  The additional test
of the values may no longer be neccessary because if the median rms
should be zero when all the values are the same.

Add commentary when the median rms is invalid.

Also space a couple of related lines to improve readability.