heterodyne: Add absolute threshold option to receptor non-linearity.
Some of the early GBS data have many and large non-linearities. Using
kappa-sigma clipping can leave receptors with non-linearity
coefficients of 0.2, when the goal is more like <0.05 and certainly
below 0.1.  This change provides an alternative fixed threshold to
eliminate all wobbly baselines above this level.  The default may be
too conservative at 0.10.  I'm using 0.08 for GBS.  Below that it's a
trade-off of losing data and signal-to-noise versus some subtle striping
in the reduced cubes.

Create a new recipe parameter for this BASELINE_LINEARITY_RMS_THRESHOLD.