SCUBA2: Add primitive to run CALCQU
This primitive was written to automate the creation of
Q and U images from all of the POL-2 commissioning
data.  It is being added to the SCUBA2 primitives
directory as it may be useful later in constructing
a pipeline for POL-2.  Currently it divides the
files by subarray and writes separate container
files for each, but this behaviour might not always
be necessary.