SCUBA2: Check outfiles list from REMSKY
REMSKY filters out some subscans so the primitive needs
to retrieve the actual list of files written, rather than
assuming that there will be an output file corresponding
to each input file.  The headers must then be read again
since the number of files in the frame changes.

A problem remaining with this primitive, and _FLATFIELD_
seems to do the same, is that when REMSKY skips over
a file in the input file list, it does not skip the
corresponding entry in the output file list.  So when it
is run with a list containing files from a number of
subarrays, some of the output files are written with
filenames suggesting that they are from the wrong
subarray.