SCUBA2: update conditions for using existing mathced filtered image
If the current Frame/Group contains multiple files, there is no
guarantee that an existing matched-filtered image corresponds to the
current file. Therefore check that the object contains only a single
file before attempting to use it.

Note that the nfiles check must be done *before* the tagretrieve test
because if tagretrieve returns true, nfiles will be 1 (as the object
now contains a single matched filtered image) and the condition for
not applying a matched filter will appear to have been met.

Also give the image name explicitly to the matched filter primitive.