general: Mask sections along a single axis.
I wanted a primitive to mask a list of emission regions.
heterodyne/_MASK_BASELINE_REGIONS_ only worked for the third axis of a
cube.  It's not clear how it worked in the context it was using of
velocitiies with pixel co-ordinates along the spatial axes.

This new primitive attempts to be more general in both the
dimensionality of the data and the axis to mask.  Hence I put in
general/ rather than heterodyne/.  It essentially creates RECT ARD
specifications and invokes ARDMASK.  Note that at present it requires
the referenced ranges to be specified in pixel co-ordinates.  Masking
can be within or beyond the supllied ranges.