heterodyne: Permits a list of regions to exclude from non-linearity tests.
Andrew Rigby requested this feature, something I had considered doing
anyway for user processing in conjunction with the automatic detection
of the emission needed at CADC.

It calls the new _MASK_SINGLE_AXIS_REGIONS_ primitive (af22a82200d7).

Although it is no long a single linewidth, the LINEWIDTH name in the
recipe parameter and argument  is retained for compatibility with
existing recipe-parameter files and invocations in recipes.