SCUBA2: store FCF in FITS header of calibrated file
There are times when it's useful to know exactly what value was used
to calibrate the data. Offline, a user is undoubtedly interested and
there are PICARD recipes that require uncalibrated data that could use
that info to undo the calibration.

Of course there is the issue of what value should be written in the
header of a Group coadd if the input files were calibrated with
different FCFs.