heterodyne: Primitive to locate the emission in a heterodyne spectrum.
It's intended to locate the emission automatically and exclude it from
non-linear baseline checks.

The hard part is to locate the emission in most cases, in the presence
of noise and even a bad baseline.  Not all the various filters are
needed for all data, but have helped with selected cases.  It may well
need more tuning after exposure to many more cases.  It's probably
adequate to exclude emission before detecting non-linear baselines.