heterodyne: New primitive to locate emission in a time-series cube.
It collapses over time and receptor, using the median for that to
reduce the effect of bad baselines, to form a spectrum.  It then calls
the new _LOCATE_SPECTRUM_EMISSION_APPROX_ to determine where the
emission lies and tags a mask of that emission for use later in
detection of non-linear baselines.

It works through each-subband independently for hybrid data.  It may
be better to re-order recipes to merge subbands before non-linear
baseline filtering.  At present you can reject a spectrum in one
subband whilst retaining it in another.  I thought I'd better store
the current state first before the changes become too convoluted,
and then rework it if hybrid merging comes earlier in recipes. The
new emission-detection primitives already have had several iterations.