scuba2: use CSO tau for extinction correction
Pointing and focus need maps to be made as quickly as
possible. Calculating the extinction from the WVM sometimes takes a
long time which can be fatal for focus observations. Makemap no longer
recalculates the WVM tau for focus observations; the current change
means that the QL pipeline will only use the CSO tau for pointing and
focus observations.