heterodyne: Resolve conflicting functionality for the AXIS argument.
It looks as if the AXIS argument was either withdrawn given the note
that the spectral axis had to be the first, or not fully implemented.
This change offers a more-flexible approach allowing not just cubes
and the spectral axis to be the first or third (time-series or
spectral cubes).  Further I take Argument AXIS to mean the spectral
axis to retain, not to trim.

Hybrid observations are merged before spectral cubes are formed,
so there is no need to pass the AXIS to the hybrid variant of this
primitive.

This primitive came under the radar because I wanted to apply it to
a representative spectrum of a time-series cube.