heterodyne: Adjust threshold usage.
This was originally intended to commit just the changes to set a higher
threshold to match the normal 0.08 non-linearity limit, but despite
various attempts I couldn't split the hunks to git's satisfaction.
Thus it now also includes the modification to bypass measuring the
histogram peak if too few values are present to form the histogram.
The limits are somewhat arbitrary.  In this case use the original
THRESH numbers and skip over the second thresholding.