SCUBA2: check product and S/N before two-component fit
Two main updates here:

 * a two-component fit should not be carried out on a matched-filtered image;
 * for "low" signal-to-noise ratios (currently <100), a single component fit is more   reliable.

The main reason for the first is that usually the fit to the
matched-filtered image takes place when determining the BEAMMATCH FCF,
for which only a good estimate of the peak flux is necessary, rather
than a fit to the profile.

For the second point, I noticed the pipeline was producing a lot of
PDA__FICMX errors, where beamfit could not estimate the uncertainties
in the derived parameters. In all these cases the peak signal-to-noise
ratio was less than about 100, a value which is too low to be able to
reliably constrain a second gaussian component with an amplitude of a
few per cent.