heterodyne: Restructure to cope with subbands and subscans.
Reinstate lost changes to use the subFrms based upon SUBSYSNR.
Access and store the spectral masks for each subband first to avoid
issues changing Frames back and forth in the main loop through the
subscans.  Add a subscan loop, which means most of the changes are
indentation of the main processing block.