SCUBA2: update default FTS-2 parameters.
Since the FTS_STAGE_CENTER position parameter is now read from
a FITS header set by the JCMT translator, and since the translator
center position was changed from 225 to 228 on May the 2nd to
make SED scans more symmetric, the ZPD position is now expected
to be around 0.0mm.  The change to the default value of the
FTS_ZPD_ESTIMATE parameter reflects this.

This in turn seems to require a change to the default value of
FTS_STAGE_CENTER so that data taken prior to the introduction of the
FTS_CNTR FITS header can be reduced with the new default value of
the FTS_ZPD_ESTIMATE parameter.