SCUBA2: crop image to 150 arcsec, not 90
If an image is going to be used for FCF calculations, then it should
be at least two arcminutes in diameter to allow for the sky annulus in
the aperture photometry; 90 arcsec is simply too small. Since the
source may not be exactly at the centre, allow some leeway and make
the image 150 arcsec instead.