SCUBA2: not much point running a source finder on a calibrator
Partly because we already know where it is, but also because running a
source finder on a bright calibrator designed to pick up 5-sigma
clumps can also pick up bumps in the error beam. Then there's the
issue that the S/N map for a bright source doesn't always have a
single peak due to higher variance associated with high signal levels.