SCUBA2: no need to apply flatfield if given all data
Calibrator maps are made using all of the available data so the
current data should always be flatfielded before combing with the
existing flatfielded data. However, in the case where the summit
pipeline processes all of the data at once there is no need to
pre-flatfield the data before running the map-maker.