SCUBA2: don't warn for calibrators in the SUMMIT pipeline
The warning about getting into a running average with no Group file
does not apply to observations of calibrators processed with the
SUMMIT pipeline. The error was being triggered by the fact that the
pipeline waits until it has all the necessary data before making a
map, so the Group ends up with multiple members without having yet
created a Group file. In the SUMMIT pipeline all those Group members
contain flatfielded data to be given to makemap.