lib: add header override support to ORAC::BaseFile
The BaseFile module can now read a header override file which
can be used if headers in the raw data are missing or incorrect.
At the start of the calc_orac_headers subroutine it checks
whether there are header overrides for the current file,
and if there are, applies them before moving on to the
header translation stage.