heterodyne: Cope with hybrid spectra that are already merged.
The extra code added in commits such as b8ec38866730 using the subfrms
method was unfamiliar because this technique is not widely used; the
merging of hybrid spectra comes early in the recipes, but I had added
the various filters in a too-advanced location before I had encountered
hybrid data.

It seems better to incorporate the full spectrum, lest only some
subbands are rejected within a spectrum.  Also the full spectrum gives
better definition of the emission.  However, leave in the subfrms code
to offer flexibility and to continue working with older versions of
recipes where the merging is performed after bad-spectrum filtering.