heterodyne: Cope better with wide wings in the histogram.
Introduce two options to prevent bin sizes that are too wide to obtain
a decent estimate of the histogram peak.

a) Use skewness and kurtosis to judge if there are extended wings,
and then evaluate the Freeman & Diaconis formula to set the bin width.

b) If the peak bin contains more than 15% of the sample size, decrease
the bin width with a fudge factor allowing for the drop off around the
peak.