heterodyne: Offset by one histogram bin.
I've long felt that the peak co-ordinate returned bty FINDCLUMPS is
displaced.  It appeared as a full pixel but a t first I could only
conceive of a half a pixel shift arising from the co-ordinate
associated with a bin being at the centre not the lower limit, but now
looking again at histograms where the peak bin is well defined plotted
in pixel co-ordinates, FINDCLUMPS position is consistently off by one
bin.