SCUBA2: do not store the refimage after all
The problem with storing the current refimage in the Group uhdr is
that it's a focal-plane reference image, which can be picked up later
at the map-making and/or mosaicking stage. This leads to wcsmosaic
errors because the coordinate systems are wrong.

So, until I figure out how to do it properly, I'll not do it at all.