heterodyne: Update the Frm files outside of the main loop.
Following a recent change (b8ec388667305) to restructure to cope with
both subbands and subscans, this primitive was losing the updated
files in Frm at the end because this Frm was not the input Frame.  Thus
the processing might as well not happened.  Now save the output files
in an array and update the Frm outside of the main loop.