heterodyne:  Expand and rework _FIND_HISTOGRAM_PEAK_.
After much experimentation this commit contains all the ideas that
survived in a sticky lump.  It's now too tricky to disentangle. Some
others may not be necessary given later work.  There are are more ad
hoc solutions and settings than I should like, and further adjustments
could be needed in the light of processing varied data.

- Rather than a CENTROID boolean argument, extend via a string METHOD
to offer more choice.  This includes Gaussian fitting.

- Add iterations to obtain better resolution and to exclude outliers
which bias the returned statistics.  Use the clump parameters to
home in on the main peak.

- Start with the Freedman & Diaconis resolution and double it on the
second iteration.

- Set the smoothing FWHM to two pixels and three pixels on the first
and second iterations respectively.

- Remove the broad background on the first iteration.

- Allow for skewness in the histogram whn trimming the histogram bounds
for the second iteration.

- Extended the prologue description.