SCUBA2: explicitly disable EXT model for focus observations
While makemap does not recalculate the WVM correction for focus
observations, the EXT model is still present in the current
pointing/focus config files. An issue with the WVM caching leads to
AST invalid object pointer errors if the EXT model is present which
would lead to further downstream errors when processing non-focus
observations in the pipeline. The solution is to disable to EXT model
altogether, which was actually planned but never implemented as the
bug didn't surface immediately.