PICARD: new recipe and primitives for comparing RMS and NEFD
The recipe can be run on processed SCUBA-2 data. It looks up the
nightly QL log file for NEPs and reads that to determine the average
value for the current observation. RMS and NEFD values are determined
from the input image, and the SCUBA-2 integration time calculator is
used to derive values for comparision. Results are written out to a
log file called log.checkrms.

It should only be run at JAC or on pueo at the JCMT as it relies on
archived log files.

Note the recipe should not be used on output from the SUMMIT pipeline
as the elapsed time and transmission values will be inorrect.