SCUBA2: allow user to force a two-component fit
By default the pipeline will now fit two superposed gaussians for
calibrator observations. By specifying the recipe parameter
FIT_GAUSSIAN=2 the user can force a two-component fit for
non-calibrator observations. Alternatively, it can be used to force a
single-component fit for calibrators as anything other than a value of
2 will be interpreted to mean a single component. This only applies if
the FIT_GAUSSIAN parameter is specified: if not specified, the default
pipeline/PICARD behaviour will apply.

Currently, the S/N threshold of 100 must still be satisfied before two
components will be fitted.

Note that it is not possible to force a two-component NON-gaussian
fit. That seems like too many parameters to adequately constrain.