heterodyne: Protect against FINDCLUMP failures to locate histogram peak.
FINDCLUMPS has been observed to fail during testing.  I had left this
in place while adjusting the recipe as it would be more evident that
something has gone wrong.  Given that now the current arrangement is
working reasonably well incorporate a status check for FINDCLUMPS too
defaulting to zero clumps.  Improvements elsewhere in the primitive
have lowered the chance of a crash, but it is always a potential risk.
In particular the clump parameter settings such as contour base level
and increment, and minimum number of pixels do need some more
consideration in the light of anomalous or pathological data, or
indeed the size of the sample.  Until I have time to look at this
again it's prudent to offer protection against FINDCLUMPS yielding an
error status.