heterodyne: Attempt to filter noisy spectra from short time series.
This addresses Tim's 19951126 data which had a single anomalous
spectrum amongst the 59 time series.  In an effort to provide a
sensible noise level above which to threshold noisy spectra I had
imposed a limitation on the minimum number of spectra to compare in
order to make FIXSTEPS and FINDBACK work satisfactorily (100 for a
50-pixel wide filter).  The lower limit of 7 (14 spectra) is a number
plucked out of the air.  We might well revise this after processing
more DAS data.