heterodyne: Don't collapse to obtain the spreads of the spreads.
The noise signal comes from comparing the spectra over time. Some of
the signal will be due to sources, but the end of spectrum noise will
dominate.  However, if the noise spectrum per receptor is collapsed
using sigma and there is a bad receptor present it can dominate the
signal.  This change selects the clipped mean for the collapse along
receptors to eject an anomalous receptor or two. There is also a note
on how this code could be developed to do this better should an image
be supplied, when it is needed.  The clipped median is also a possible
estimator for collapsing the receptor axis.  Experiments on hybrid
data with two bad receptora of only twelve, the clipped mean gave good
results.