lib: FLUXES no longer needs shenanigans
Previously we needed to set FLUXES and FLUXPWD and ensure
that a JPLEPH soft link existed in the working directory. That
was all removed in a commit to the Starlink tree back in August
2012 and is no longer needed. Since JPL is no longer part of the
Starlink build the code that relies on JPLEPH needs to be removed.
Fluxes can simply run directly now.