heterodyne: Extract common code for obtaining and reporting recipe parameters.
Start with three which cover the high- and low-frequency interference,
non-linear receptors, and flat fielding.

Allow additional releated parameters to be set without the overarching
flag parameter.  In three of the four cases the flag defaults to true
in the primitive that does the desired step. It seemed silly to force
the user to add the boolean selection parameter in such cases if they
only wanted to refine an associated parameter.  Also add the heading
message should an associated parameter be set without the selection
parameter.

Allow for an existing true setting to uhdr INDEX_BAD_RECEPTORS now the
code is fragmented into primitives.