SCUBA2: delete pointing flag file after creation of new map
Deleting an existing flag file at the beginning of the recipe means
that in the case where the pipeline hangs, a pointing solution will
not be estimated even if a suitable file exists. Therefore, only
delete the flag file once a new map has been created, thereby reducing
the likelihood that an existing file will not be used to derive a
pointing solution.

Of course, this doesn't help those cases where the pipeline hangs
processing the first subscan.