SCUBA2: store actual new bounds in uhdr but for makemap only
Previously this primitive only stored the flag to indicate whether the
bounds check had been successful. It would be more useful if the
actual bounds could be retrieved - and so the new bounds are now
stored in a hash reference under the NEWBOUNDS uhdr.

Note the storage and retrieval is only done for makemap calls:
wcsmosaic calls will always re-calcaulte the bounds as
mosaicking/coadding may involve fields offset from one
another. Results for makemap will always apply to the same field.

This commit solves an occasional problem in the SUMMIT pipeline in
which new bounds are calculated at the map-making stage but not for
the mosaicking stage because the Frame uhdr NEWBOUNDS was
defined. However, the pipeline would then crash when the caller tried
to dereference the non-existent LBOUND entry in the primitive hash.

Most of this commit is to get the indentation correct which I should
have done when I first added the check for existing bounds.