SCUBA2: no gaussian fits for matched-filtered calibrators
While it's not always the case, fitting a matched-filtered calibrator
image is usually done as part of determining FCFs, for which the
requirement is that the peak be fitted accurately, and the profile is
less important. Therefore allow the radial falloff to be a free
parameter to allow beamfit to make its best possible fit, unless the
user has specified that a gaussian should be used.