heterodyne: Add DILATE argument to high-frequency noise filter.
When ringing is present the actual noise used for thresholding is
enlarged by weak ringing signal away from and strong ringing band.
As a consequence the threshold to remove the spectra affected by
strong high-frequency interference is cut too high, and this leaves
one or two such spectra at the periphery of bands as the inteference
grew or died.

The downside is that it may lose a few good spectra if there are
isolated single spectra with interference.  The alternative is to
decreaed the default clipping factor when HIGHFREQ_RINGING is
selected.