PICARD: trim map coadds and look for 4-sigma sources
To aid the source finder, the maps should be trimmed to exclude the
noisy edges even though the search is performed on the S/N map. In
this case a circle of radius 4500 arcsec seems to do the trick.

Also lower the detection threshold to 4-sigma to see what that brings
(I've been running it this way offline for months anyway).