SCUBA-2: add direction-specific ZPD calibration
This commit adds a recipe parameter FTS_SCANDIR_ZPD which
can be used to switch on direction-specific ZPD calibration.
By default this is currently off, retaining the previous
behavior.

When direction specific ZPD calibration is enabled, the ZPD
position frames are stacked by SCANDIR as well as subarray,
making use of the fact that FTS2SPLIT already helpfully
corrects the SCANDIR FITS headers.  Provided the ZPD observation
comprised more than one scan, this will lead to two ZPD files
being produced per observation.  They are logged in the
calibration system as "zeropath_fwd" and "zeropath_bck"
since it is necessary to handle both directions within a
single observation.  Then _FTS2_INIT_ will look for the
direction-specific index files.  This means that it is
necessary to turn on FTS_SCANDIR_ZPD for either both
REDUCE_FTS_ZPD and REDUCE_FTS_SCAN, or neither of them,
in order for the ZPD calibrations to be found.