heterodyne: Only find the emission map if it going to be used.
Yes this does add a dependency knowing what else will use the
approximate emission map, but it seems silly that it's being run when
its results will not be used.  When we come to iterate and use
the identified clumps to mask the input time series for non-linearity,
a further test will be needed.