heterodyne: Fix wrong rsp for stares/jiggles
Altered _CREATE_MOMENTS_MAPS_THROUGH_SMOOTHING_ for jiggle and grid
observations. It now first of all tries the 1,1 pixel position to
extract the representative spectra, and if that is blank it uses the
unsmoothed peak position. This should fix the bug where the smoothing
used to avoid noise spikes causes an empty pixel to be used for some
stare modes (scicom trac 96). Not sure if using it for the jiggle
observations is a good idea though...