PICARD: crop images to 90 arcsec, delete temp files
Some updates requested by users - the input maps are now trimmed to a
circle of radius 90 arcsec before analyzing. Documentation for
changing this value via a recipe parameter has been added.

Intermediate files are now deleted.

The documentation for the ITC_F parameter has been removed as it is
calculated automatically as part of the analysis to ensure
self-consistency.