heterodyne: Smooth the noise profile before analysis by MFITTREND.
This is to prevent there being too many sections (>20) of excess noise
for MFITTREND to handle.  It's the broad region of excess noise at the
ends that need to be removed.

The smoothing box is a number picked out of the air.  I tried 15 first
and that worked but still left weak ripples.