heterodyne: Merge hybrid observations before filtering bad spectra.
This has the advantage of having uniform thresholds across the
sub-bands, not rejecting parts of a spectrum, better automatic
emission masking, and normally a better handle on the baselines.  It
should be faster too. This order works much better for DAS->ACSIS
data.

This has been pending a couple of months, as the original tests were
on the staff time where the slopes of the baselines were different in
each sub-band.  That appears to be an exceptional dataset being of the
Galactic Centre with a very wide velocity range and difficulties
obtaining a clear reference position.  If the differing baseline
slopes is an inherent ACSIS issue, it may matter little as such hybrid
observations demanding the extended spectral range are uncommon in my
experience.