SCUBA2: primitives to track observation properties to compare with ITC
This trio of primitives form the basis of retrieving and logging
so-called "CHECK RMS" data, named after the PICARD recipe. In
practice, the aim is to store observational parameters along with
noise and NEFD values for processed images in order to compare with
the expectations from the integration time calculator (ITC).

Each primitive represents a particular step in the collection of these
data.

 * _GET_CHECKRMS_DATA_FROM_HEADER_ retrieves all the necessary
   information from the current Frame/Group header.

 * _GET_CHECKRMS_DATA_FROM_IMAGE_ retrieves the exposure time, noise
   and NEFD data from the current Frame/Group image (within a given
   aperture, default is 90" radius) and obtains the corresponding
   values from the ITC.

 * _WRITE_CHECKRMS_LOGFILE_ writes a log file (log.checkrms) with all
   the desired results.